Fix: mkstairs did not treat grass as acceptable for stair placement
When I got this impossible from it attempting to place stairs on grass,
I saw that the code is doing one of those "terrain type is an 'ordinary'
walkable space (room, corr, ice)" checks which simply didn't account for
grass. So add it there; it is supposed to be fine to transform any grass
space into stairs.