Prevent Gehennom demons generated in mklev from usually being asleep
This logic of starting 80% of demons created with the level as asleep
goes back farther than NetHack has git history for, and my best guess is
that it is based on assumptions that most of them would be in a tightly
constrained maze or demon lord lair where it was desirable for them not
to wander around randomly and set off traps. If this is correct, that
logic doesn't carry over now that Gehennom is mostly free of mazes. The
default behavior should be for demons to be awake. (If trap triggering
in lairs is a problem, those monsters should be set to asleep=1 or
waiting=1 manually in the lua file.)

This does not affect demons in graveyards (always asleep) or any demon
that ever happens to generate or be placed outside of Gehennom in the
absence of other things specifying its sleep state (which will retain
the 80% asleep logic). So it would apply to non-graveyard demons in the
Priest and possibly other quests, but we'll cross that bridge when we
get there.