Prevent items from randomly generating as silver in Gehennom
There have always been odd reports like "the marilith is attacking me
with a silver sword" (that their weapon randomly generated as its object
material), which don't make any flavor sense. I figure that it's
reasonable that the mostly silver-fearing inhabitants of Gehennom would
have either kept out or systematically removed any silver items that
happened to be there. So, now the only silver found in Gehennom is what
you bring there yourself (or items that happen to be silver by default).

At first this applied only to level creation, but it doesn't really make
much sense for death drops to be silver either, or for monsters spawned
after level creation time to carry silver items. So now it just blanket
applies within Gehennom. Note that things which could force the material
to be silver (i.e. a wizard mode wish) still work, since they apply the
silver material after mkobj has finished and selected a material.

I considered adding code to force silver-base items which could possibly
generate as some other material to do so, but decided not to at this
time, so shields of reflection and such can still be silver in Gehennom.