Implement the Stygian Marsh level
This is the fifth main-branch Gehennom level, coming immediately before
Dis. It serves a couple structural purposes: the branching off point to
Shedaklah (swampy and Juiblex-containing) and the visible outskirts of
Dis.

As a swamp, it's pretty easily traversed, and there aren't too many
enemies to fight if you just want to pass through. Beneath the water,
though, there is some treasure, and a large number of unrotting corpses,
semi-inspired by Tolkien's Dead Marshes. They do not take kindly to
their treasure (which includes anything that happens to be dropped in)
being disturbed.

Added a few related rumors as well, and made a small adjustment to the
swamp in the Dis level so it matches this terrain better.