Slightly tweak the Juiblex level
I can't really think of many edits to make here. Juiblex's swamp is more
or less already the level I want, with more or less the correct
monsters, objects, and traps. It becomes a *little* less distinct with
the addition of the Stygian Marsh, but as a demon lord lair it's fine.
The main changes I'm planning will be with Juiblex itself.

I allowed the level to flip horizontally, added a couple rust traps, and
a few additional monsters. (This is the first of the four existing demon
lord levels to be edited in the Gehennom overhaul, and the other three
will follow the same philosophy that the overall difficulty doesn't
necessarily need to be kept the same since it is no longer necessary to
complete or even enter the level to complete the game.)