Track the number of Juiblexes anywhere in the dungeon
This is necessary because I later want to check whether a demon lord has
been killed. For most of them this is as simple as checking
mvitals.died, but for Juiblex this will be true as soon as you kill one
of his clones and haven't actually eradicated him. Furthermore, killing
the last one of him on a level doesn't mean you killed all of him,
because one or more clones could have gone to another level. So he needs
a Rodney-like counter to track how many of him currently exist.