Fix: door specified only as "closed" could end up locked
Due to the door generation code for lua files assuming that anything
unspecified should be randomized. But in fact it's conventional that
specifying a door as "closed" and not "locked" means the door should not
be locked. This led to shops in Minetown being incorrectly closed for
inventory (they should always be open, except Izchak in Orcish Town),
and probably lots of other special level doors throughout the game.

Fix this with logic that if the door is specified as closed and locked
is left unspecified, it should never be locked.