Refactor: replace monster vs trap logic with immune_to_trap
immune_to_trap was added all the way back in xNetHack 1.0 as a function
that would work equally well on the player and monsters but for some
reason, it was only ever called once for testing if the player should be
immune to a trap.

Now, add another call in mfndpos, replacing a long and less detailed
conditional for the monster movement case, ensuring that the same
decision paths will be followed. (Now, for instance, a demon will
consider whether it's carrying burnable items and not just its own
immunity to fire before deciding to move onto a fire trap.)