Remove fakewiz1 and fakewiz2 levels
These already did not generate in the Gehennom structure anymore, but
were still being added to nhdat and shipped with the rest of the game.
Since they are not going to be making a comeback, remove them
completely, as well as Is_portal_level() which was used to look up the
one with the magic portal.