Add the cold trap
The Cocytus branch is going to take some cues from UnNetHack's icy Sheol
branch, and this is one of them - a branch-exclusive trap that causes
cold damage, similar but opposite to fire traps in the rest of Gehennom.

Cold traps differ from UnNetHack in that they deal substantially more
damage and can strip intrinsic cold resistance (in an attempt to solve
the problem of them being a total non-threat for anyone at the point of
reaching Cocytus). The flavor is not a blast of cold from the ground,
but instead a (probably magical) mechanism sucking heat from whatever is
above it. Combined with other planned Cocytus hazards, having extrinsic
cold resistance is therefore more important.

Beyond this trap, there are no plans in place to port other Sheol
features (blue slime, ice golem, ice walls, etc).