Cocytus causes persistent cold damage if hero is not cold resistant
This is adapted from similar rules in the Ice Queen's Realm in EvilHack.
Once per turn (not action), the hero will take 1d6 cold damage while in
the branch, if they are not cold resistant.

I had to rework EvilHack's code a bit, sacrificing some death-message
quality, because calling losehp() with its unavoidable "You die..."
message is preferable to reimplementing the guts of losehp() to
correctly handle if you freeze to death while polymorphed. (In EvilHack,
this is currently a bug -- freezing to death while polymorphed kills
your normal form without warning.)

As the comment in this commit explains, the cold damage does not happen
to monsters. While I don't like this asymmetry on principle, it will
make it much easier to put monsters in the special levels without
worrying about a cheesy strategy of arriving on the level and waiting
for the non-cold-resistant monsters to die.

Also, moving into and out of Cocytus now produces the correct messages
regarding the heat and smoke of Gehennom. Leaving Cocytus via
levelporting or branchporting to a non-hellish destination does not
produce any message, which is consistent with the rest of Gehennom.
(There isn't a "heat and smoke are gone" message if you leave via
levelport or branchport.)