Limit monster generation in Cocytus to demons and cold-resistant
This accomplishes a few things:
- Takes care of flavor issues like "why is there a [obviously not
  cold-amenable monster] appearing in Cocytus?"
- Partially justifies the decision not to have non-cold-resistant
  monsters lose HP from the cold, because now most of them would not be
  losing HP anyway
- Puts the player at the biggest risk from cold traps
- Allows demons to generate like normal; without this exception, the
  only demons that would ever generate naturally are ice devils.

I considered an implementation involving a new makemon flag named
something like MM_RESPECT_ENV which would move this cocytus check into
makemon and be passed in for the over-time monster generation mechanism.
The benefit of this would be that monsters generated with the level, and
naturally over time, would be okay with the cold, but summoned monsters
would not necessarily be. However, since non-cold-resistant monsters
won't lose HP, that's not as important to have. (And anyway the status
quo is that create monster will respect the local generation rules --
e.g. summons in the Quest respect Quest generation.)