Foocubus encounters grant temporary cold resistance
The primary reason for this is to open up an interesting option in
Cocytus: if there is a foocubus around, you can get cold resistance from
them by agreeing to their seduction, which could be life-saving
depending on the situation. However, it might also be situationally
useful outside Cocytus as well.

Prior to this, it wasn't possible to get temporary cold resistance.
There was no message for it timing out, which I think is correct to
leave as it was. After all, removing the ring of cold resistance does
not produce any messages, so there should not be any perceptible
difference there.