Redistribute wands of wishing through Gehennom
The overall number of guaranteed wishes is the same - 6 - but they
appear in different places. Notably, the Castle only has one now, and 3
are given to random demon lords. (The other two remain with Vlad and
the Wizard of Yendor.)

The motivation for this is to provide a soft motivation to go after
demon lords rather than ignoring them. (There is another soft motivation
planned, which is not yet implemented.) I have been debating whether to
include a hard motivation, of the "the invocation will fail if you
haven't dealt with at least N demon lords" sort, but at the current time
I think it will make for better gameplay without that.

In order to guarantee that exactly 3 demon lords will get a wand of
wishing, I implemented a system that randomly picks 3 at the start of
the game and sticks this in global state. Combined with a Lua accessor
that checks whether a given demon lord is present in this state, a
demon lord lair file can now check whether it should generate a wand of
wishing or not. (Interesting follow up idea: perhaps the Oracle could
use this information to give some clues about where to look for the
wands.)