Implement Asmodeus's lair
At the bottom of Cocytus, this is a fortress in a desolate, icy cavern.
Devils, and coldness abound. As with subsequent demon lord lairs, this
one uses the feature from the last commit to determine if there should
be a wand of wishing or not. The general shape draws from the former
design - a long hallway leading to a square fortress shape, only more
elaborate.

One new thing about this level (that will probably end up being used for
at least one other demon lord lair) is that Asmodeus is not generated as
part of the level. Instead, an unfilled throne room is designated at the
heart of the fortress, and entering this area causes him to appear (with
none of the preexisting weird perm-invis shenanigans going on in the
back end). I like this because it provides an unambiguous point at which
to call boss_entrance, avoiding the various annoying problems where it
gets called at the wrong time (after a bribe, a couple turns into the
fight, etc.)