MS_BRIBE demon lords demand valuables for safe passage
This expands the bribe system with a feature from SpliceHack, though
not taken exactly. The way things work now is that a demon lord will
roll 50d1000 for the amount of gold they are going to demand as a bribe,
then inspect your inventory for valuable items, and demand random ones.
If you consent to hand them over, they take it and deduct its (inflated)
value from the amount of gold they're still planning to demand. They may
then ask for more items, or just ask for the balance of the gold.
Appeasing them completely means they will poof away along with any items
they have taken. Refusing to hand them any item gets them angry just
like if you refuse to pay them gold.

The idea behind this is to make the player face some tougher choices
than "do you want to give up the gold you've collected which is more or
less useless for anything else at this point in the game". Are you
willing to part with a valuable artifact, bag of holding, or
wish-granting item that you're carrying? Not as easy a choice.

I don't think paranoid confirmation is needed for the "Give up your
item?" prompt because if you hand over something you didn't want, you
will almost certainly be able to refuse to give the demon their next
demand, unless it's the last in a long string of high-value items.