Add "doomed" intrinsic
This is inspired by a similar feature from ADOM. Unlike in ADOM, though,
being doomed does not have a huge laundry list of things it specifically
affects negatively (for now, at least). Instead it simply treats most
things that check your Luck as if your Luck was -10.

One of the main effects of this is that the player will suddenly find it
much harder to hit things, though there are also a bunch of side effects
that come with bad luck: triggering more and worse traps, searching,
writing unknown scrolls, etc. (One thing to note is that I'm considering
removing Luck from to-hit calculations. If it turns out that that's the
main drawback to being doomed, then probably other side effects of the
condition should be added.)

This is intended to be a pretty nasty effect that you can only get from
tangling with one (or two?) specific demon princes. Though it would be
intriguing to have e.g. some artifact that doomed you while you wielded
or wore or carried it.

Since Doomed is implemented as an edit to the Luck macro, a couple
things that checked u.uluck (base luck) directly needed to account for
dooming. This includes getting good effects from magic fountains and
getting a sacrifice gift.

Very minor side effect of this change: the computation for whether a
cloned monster created by the player would retain any tameness or
peacefulness checked u.uluck directly, and I found it most expedient to
change that to a generic Luck check. This means that when not doomed,
luckstones will influence these odds, but this happens rarely enough
that I doubt anyone would notice anyway.