Add persistent intrinsic blindness
This is not actually a new intrinsic, of course. It uses the existing
BLINDED property. But up until now, the FROMOUTSIDE field of that
property (indicating blindness bestowed on the player from an external
source) was not used. This commit implements the ability for it to be
used; nothing causes intrinsic blindness yet but something will in an
upcoming commit.

"Persistent intrinsic blindness" behaves rather more stickily than
temporary, timing-out blindness, acting somewhat similar to roleplaying
blindness. Most importantly, make_blinded(0) calls do not remove this
condition, which means that carrots, unicorn horns, healing potions, and
the like are not effective when the hero has been inflicted with it.

Four things can remove it: the spell of cure blindness, prayer, the
Staff of Aesculapius, and the Eyes of the Overworld. This is meant to be
tricky enough that you can't trivially remove it with gear every
adventurer carries, but not something you're supposed to live with for
the rest of the game if you don't want to. And as a bonus, cure
blindness now has a purpose again.

Prior to this, "Blinded" was assumed to always represent only a timeout,
and thus setting it directly was something of a common pattern, e.g.
"Blinded = 1L;". Now, since statements like that will strip intrinsic
blindness, I had to rework many of them so that they only modify the
timeout, and leave the intrinsic alone. Also, it was _probably_ safe to
leave statements like "make_blinded(Blinded + foo, xxxx);" alone -
because it will call set_itimeout and the actual timeout change won't
affect anything outside the TIMEOUT range - but I changed them all to
"(Blinded & TIMEOUT) + foo" to be safe. The former convention should no
longer be used in xNetHack.