Fix #103: dismounting steed over open air/chasm
dismount_steed() was failing to check that the monster was grounded
before calling mon_aireffects() and consigning it to splat on the
distant floor.  This, combined with landing_spot() using Flying (which
the hero's flying steed contributes to) as reason to consider open air
as a safe dismount position, caused some potentially-deadly problems
(both for the hero and her steed) when dismounting over a chasm.  Should
fix #103.