Revamp Asmodeus's spellcasting
This is all that is currently planned for him: no especially fancy fight
changes, just elevating his magic attacks to the nastiness worthy of a
demon prince.

He now casts spells from a special list like Dispater, but his list only
contains 3 spells. The heavy hitter here is Dark Speech, a feature which
originated in pre-rewrite SpliceHack as an ability for any demon lord.
As a special spell for Asmodeus, I modified dark speech to be nastier
than it already was. The idea here is to provide one of a random
assortment of immediately bad effects and long-term bad effects, as
well as hit the hero where it hurts (at this stage of the game, HP is
cheap). Thus, effects like applying intrinsic blindness and doom and
stripping intrinsics for the long-term detriments, and withering and
psychic damage with short-term combat implications.

The other powerful magic he wields is his signature cold spell, which I
extracted into its own arcane monster spell because he no longer has an
AT_MAGC AD_COLD attack. This is buffed in two ways: cold resistance
reduces it to 25% instead of fully negating it, and it can freeze
potions in open inventory (which is kind of a no-brainer; why didn't it
do this in vanilla again?)

I also made Asmodeus a non-warper. This might end up making him too easy
if one stays at range, but also if you stay at range he will cast
exclusively dark speech at you, because he can only cast sheer cold at
melee range (due to an annoying limitation of how castmu is called at
range).

He is also no longer infravisible, since his association with cold is
now underscored much more strongly than in vanilla.

Note that Asmodeus appears as peaceful unless you are wielding
Excalibur, and is still bribable. It is not the intention that everyone
should feel the need to fight him.