Fix: trap doors didn't work properly in Gehennom side branches
There was a special case in the code intended to prevent trap doors from
prematurely allowing the player into the Sanctum. But it had an implicit
assumption that there was only one hellish branch in the game. This
became a problem in Cocytus, which ought to allow falling into Asmodeus'
level, but instead trap doors just didn't work.

To solve this, I added a new macro In_main_gehennom which checks for
only this branch, which necessitated adding another case to the "look up
and save dungeon index by name", but hey, it's better than hardcoding 1
and praying that never breaks.