First entry into any throne room in a demon lord lair makes them appear
This is probably what most of the demon lords will end up doing: they
won't generate at level creation time, but will instead have their
appearance triggered by the hero entering a predefined, unstocked throne
room. This was already used for Asmodeus. It may not end up being used
for Dispater, though.

In order to accomplish this, this commit also refactors out the logic of
determining who is in charge of a level (and extends it to cover several
other cases that weren't accounted for previously). This used to be used
only by the nommap-boss logic, but now also needs to be used to
determine which demon lord to spawn in.