Implement Baalzebub's lair redesign
Unlike most of the lair redesigns, this keeps a major element from
vanilla, which is the bug-shaped fortress. Except here, it's in a cavern
and its shape is clearly visible without magic mapping. I tweaked the
legs a bit to make them more proportionate to the bug.

This is one of two levels (the other is Geryon) that has some terrain
and monster variation depending on whether it's in Cocytus or not. This
worked more smoothly than I expected.

Since it's no longer in a maze, the baalz_fixup workarounds and
g.bughack are no longer needed and I have deleted them.

Also, I moved the 'lawdemon1' lair (which could also be Geryon's) to
the middle level in Cocytus, rather than the entry level. The entry
level is now not a special level and will be cocytusfill.