Backport coordxy type from NetHack 3.7
It's not a desirable outcome to continue using xchar for coordinates
when 3.7 has moved on to a dedicated type for them, so going forward I
would like to use coordxy for variables that represent coordinates.

This is written to set up a merge conflict with vanilla, but no behavior
should be affected by taking vanilla's version at that time because the
coordxy types are the same.

In this commit, I went and changed all the new additions of xchar in the
gehennom_overhaul branch to coordxy; there are plenty of other xchars
representing coordinates that are in xNetHack and not vanilla, though,
which will be addressed when I next merge vanilla.