Add giant fly monster
xNetHack doesn't add many monsters, but this particular one is going to
be more of an enhancer for Baalzebub than its own thing. It doesn't
generate normally, and will only be generated in conjunction with him
and his level.

The details of the giant fly including its tile come from SpliceHack,
but I upped its speed, difficulty rating, and damage (1d4 to 2d8). It
retains a pitiful amount of hit dice; these are supposed to go down
pretty easy but be dangerous if you find yourself in a large swarm and
unable to dispatch them efficiently.