Add giant flies to Baalzebub's level
Which allows them to serve as an antepiece rather than having them first
be encountered when Baalzebub is fought. Plus it makes flavor sense that
flies would be hanging around here.

Also fix a problem where if this level appears in Asphodel, the stairs
to Gehennom would be randomly placed.