Remove the corrmaze system from the game
Mazes themselves have not been entirely removed yet -- they are still
barely hanging on in appearances between Medusa and the Castle, as well
as whenever a lua file fails to load -- but the variant of the maze with
corridors instead of floor and stone instead of walls only appeared in
the vanilla incarnation of Baalzebub's level, and now that it's been
changed, does not appear at all. I have no intention of ever
reintroducing it in a level, so this commit removes all the various
pieces of code that make it work.