Refactor: extract semantics of "won't leave level" into function
This is in preparation for an upcoming feature. There were already
various safeguards against shopkeepers, priests, and guards getting
off-level by quaffing cursed gain level or zapping a wand of digging or
what have you. This puts that into a function, where other cases can
easily be added in the future.

Additionally, this prevents monsters which won't leave the level from
using stairs or ladders. There is also a very small adjustment to
behavior: shopkeepers used to be able to use cursed scrolls of
teleportation (but still not wands of digging or cursed gain level, etc)
when they were not in their shop. When I was writing wont_leave_level I
had to decide whether shopkeepers should have all the off-level escapes
available once they leave the shop, or still not want to leave the
level; I decided on the latter, because otherwise they can never go back
to tending the shop. So they can no longer use cursed teleportation in
any circumstances.