Implement Geryon's lair
Asmodeus has an icy fortress. Baalzebub lives inside the giant shell of
an insect. Geryon has... a ranch. Unlike most other demon lord levels,
this one is fairly open and not difficult to navigate. Most of the space
on the level is occupied by big pastures, in which Geryon's herds of
peaceful q monsters can be found. If the player chooses to pay a bribe
to Geryon, they won't have that much difficulty getting in, making their
deal, and getting out again. I imagine Geryon as the most businesslike
of the lawful demon lords, and willing to be hospitable. (I may be
influenced here by his portrayal in the novel The Battle of the
Labyrinth.)

As with Baalzebub's lair, this one has some terrain variation depending
on whether it's in Cocytus or not.