Create a global context struct for tracking archfiend info
This is the planned successor of wish_dlords (which will be retired in
the next commit), and is also required for the final Gehennom system
I plan to implement, which will depend on tracking whether the hero has
actually dispatched a demon lord or not. This is being implemented
early because I realized I needed it to cover an edge case in Geryon's
behavior.

As of now, this tracks

- has_wish: randomly chosen at the start of the game, will replace
  wish_dlords.

- num_in_dgn: replaces no_of_juiblexes and is generalized to any
  archfiend; it is useful to be able to track if a demon lord exists
  elsewhere in the dungeon if they aren't on the current level.

- escaped: set true only when the archfiend flees the dungeon, which
  distinguishes the case of them being gone from the game because you
  bribed them, from the rather easier case of taking them to level 1 and
  coaxing them to flee up the stairs.