Make Geryon more formidable in combat
Geryon has been slow and easy to defeat for several versions of NetHack
(possibly since he was introduced). This adjusts that to make him worthy
of having his own lair and possibly guard a wand of wishing like the
other archfiends, without actually changing his statblock.

The idea implemented here where he is stronger the larger his herd is
was one of the first ones discussed online for how to make archfiend
fights feel more varied. An interesting spin I added to this was that if
he weakens because the player is thinning his herd to take down his
regeneration and damage boost, he gets faster. A boss fight where he
steadily gets weaker and less threatening until dying was not very
compelling to me.

He no longer covetous warps, except in one case: if you kill his herd
animals before killing him, he will warp to you immediately and become
angry (which isn't really a covetous warp). This happens even if he is
bribed, because the bribe is for your safe passage and his
non-interference later, not to give you free rein to attack his
creatures. Of course, if you've bribed him, there isn't really any
reason to attack his herd because you won't be fighting him.

If Geryon is present but elsewhere in the dungeon when you attack the
herd, he will not appear, so this commit also prevents him from
voluntarily leaving the level. There are still ways to get him
off-level, and everything should still work if he is, but it won't be as
easy as luring him to the stairs and having him flee up or down them.