Enable monsters with a specific class and alignment in special levels
This is to enable the specification of lawful demons in lawful demon
lord lairs, and chaotic demons in chaotic demon lord lairs, without
making specific assumptions about which monsters appear in the monster
list.

This is a bit of a hack, because it overloads the (unused, and unlikely
ever to be used) path of creating a monster of a definite class and
/in/definite species who is a minion of a specific god. The "correct"
way to do this would be to add another alignment field to the sp_lev
monster struct and treat it completely separately from the existing
"align" field, but since I don't expect that behavior to ever be needed,
it wasn't worth doing the "correct" way.

There are plenty of existing places in special levels where minions with
a specific alignment are specified, but they all have a specific species
(either aligned clerics or Angels), so no existing levels are affected.