Make ice and bone devils lawful
This solves the problem of lawful demon lords having only horned and
barbed devils to fill their lairs with, without resorting to chaotic
demons. In modern editions of D&D, these two are lawful in any case; I
think it's only in the first edition, where the chaotic = evil conflation
was prevalent, in which they are not lawful.