Convert lawful demon lord lairs to generate lawful-biased demons
Taking advantage of the recently added system that enables this. In most
cases, the demons are not /guaranteed/ to be lawful; there is a small
chance of it allowing a random alignment instead. The goal is to give a
noticeable difference in inhabitants for lawful versus chaotic levels,
but not necessarily make it strictly limited based on the demon lord's
alignment, which would be completely predictable, and predictable is
boring.

Chaotic lairs haven't been touched because 3/4 of them haven't been
implemented yet and Juiblex just has a few random demons hanging around
that don't particularly need to be chaotic biased.