Implement Yeenoghu's lair redesign
Similar to Juiblex, he doesn't really have a fortress or palace; this is
intended as merely locating the area within his wasteland where he
currently is. Because of that, he doesn't have a throne room that spawns
him in, and is just present on the level from the beginning. I also
didn't really find any good place to stick his wand of wishing, so
instead he carries it on his person in a sack.