Buff Yeenoghu in combat and make him immune to fleeing
Yeenoghu's niche is that he is the only demon lord who goes on full
attack. He will not retreat or escape. The goal of most of these
archfiend changes is so that they become distinct from each other, but
this still allows one to be a straight fight where you simply need to
overcome their overwhelming force. Yeenoghu is that one.

This wasn't described in the prior commit message, but he is now given
in the Lua file a +9 rusty flail named Butcher (not an artifact), which
is the main way his combat is buffed. He also has an extra 2d6 of
damage, and much better AC.

He is now immune to monster fleeing. I reviewed the cases where monflee
is called, and most of them either behave fine if the monster fails to
flee (i.e. not printing a message saying that, or being specific to a
circumstance that won't apply when fighting Yeenoghu).

Side effect: it was possible to apply a mirror to Yeenoghu and get a
message that he is frightened of his reflection, even though he wouldn't
actually flee. So in this commit I also made applying a mirror not work
on any major demon. They should be well acquainted with what they look
like.

Also port a typo fix from SpliceHack in his appearance message.