Implement the Abyss filler level
This is a rather minimal level because so much of it is open space, but
it still won't be particularly easy to traverse. One thing I added late
in the design was the system of demotivational engravings.

Despite the engravings, jumping into the Abyss does not result in
instadeath, but it does send you directly to the bottom level even if
there are intervening Abyss filler levels. I felt like since no one is
going to willingly jump in (unless the demotivations are a lot more
powerful than I thought), anyone falling in will be doing so by accident
or due to monster knockback, so having it be instant death would just be
cruel.

To accomplish that behavior I had to add In_abyss as a macro, which
would probably be more useful if it were a longer branch.