Implement Demogorgon's lair
Intended to be a serious challenge, to fit its seriously challenging
demon lord. The level design here reuses a concept I once had planned to
apply to all of Gehennom: special, one-time-use keys that open doors
that cannot be bypassed any other way. In the original formulation,
these would have appeared all over Gehennom and led to side areas in the
filler levels containing demons and treasure. Two things got in the way
of this: first, since levels are only generated when physically entered,
it would have been difficult to adequately forecast the balance of keys
to put on which levels; and second, it was intended for the
standard-length Gehennom with lots of filler levels, which no longer
exists in xNetHack. In a single level, though, the system can be an
interesting challenge without overstaying its welcome.

Most of the mechanics of this system are the same: a special type of key
(currently, any bone key, which allows the player to get extras of them
if they find or wish for them elsewhere) that is the only way to open
one of the level's special doors (currently, iron doors). The doors are
not supposed to be destroyable, magically openable, unlockable,
pickable, phaseable, squeeze-under-able, or any other method of
circumventing them. Just the keys which are strategically placed around
this level.

One thing I may end up changing later is to make the doors bone, not
iron (though keeping the same indestructible properties) and overloading
spe on the keys or something so that they appear as "ornate bone key"
and thus will make it easier for players to cotton on to how to use them
to explore the level.

Nothing is fixed in the level apart from the upstairs -- random room
contents, which doors will appear and what type they are, Demogorgon's
location even, so the player may stumble upon him if they aren't
prepared for it.