Allow des.object trapped field to take a boolean
Unlike all the other monster and object fields that accept only an int
(eroded, recharged, etc), the trapped field is ONLY ever useful as a
boolean, and it's unintuitive for level designers to have "trapped=true"
in an object specification error out. Change it so that it will accept
both booleans and ints.