Demogorgon buffs: withering and intelligent spellcasting
Since all the other demon lords are being buffed, Demogorgon needs one
too in order to remain the toughest. This is not an incredibly large
buff though, especially since he is no longer covetous. He gains a
withering bite (this was going to be a drain life attack, but it turns
out he already has that), and his spellcasting prefers healing when he
is at low HP and can otherwise cast dark speech (same formula as
Asmodeus, except Demogorgon uses the full array of non-special arcane
spells whereas Asmodeus only uses sheer cold).

Initially, the plan for Demogorgon fight changes was to implement a
system where he would frequently jump away and evade most attacks, only
to be unable to do so if the player had pets flanking and distracting
him. I still kind of want to do this, but didn't want to figure out how
to balance it right now.