Another Yeenoghu buff: disarming attack
Not a specific attack or new attack type. This is added to his primary
weapon attack: a 10% chance that your wielded weapon is swatted out of
your hands and sent flying in a random direction. This matches
Yeenoghu's berserking attack style fairly well.

Since I found using the scatter code to be the best way to send the
object flying, this commit also includes a refactor to the scatter flags
to allow the caller to specify a direction for the flying object. (This
is necessary so that Yeenoghu doesn't illogically swat your weapon out
of your hands and send it flying into himself.) Existing callers
shouldn't be affected.

Yeenoghu is also no longer a covetous warper. I think as of the initial
buff commit for him, he remained a warper in order to compensate for the
fact that he won't take any other action to protect himself. With this
addition I feel okay removing that compensation.