Extend and redesign Orcus's level
Not a full overhaul since much of the original level design is kept.
This shifts the ghost town part of the level over to the left side, and
uses the space freed by removing the maze to add a skull-shaped lair for
Orcus, rather than having him just hang out in a random building.

Monster distribution is more or less the same, with some extras
inhabiting the skull lair. Specifically, there is a lot of undead in
this level, whereas undead have been toned down in the other lairs.