Add percent() and rnf() random number helper functions
Both of these are intended to write more semantic code and avoid
less-readable expressions involving rn2.

rnf() [random number fraction] is from UnNetHack, and is for when the
developer wants to have something evaluate to true in fractional terms,
i.e. with a 3/7 chance.

percent() is intended to work the same as it does in Lua, for when the
developer wants to have something evaluate to true a fixed percent of
the time.